Username showing up on profile + data between tabs fix
